Clean up KernelInvokeApi
