+++ radix-engine-interface/src/api/types/invocation.rs
@@ -1,7 +1,6 @@
 use crate::api::types::*;
 use crate::blueprints::resource::{FnKey, MethodKey};
 use crate::data::scrypto::model::*;
-use crate::data::scrypto::ScryptoValue;
 use crate::*;
 use sbor::rust::prelude::*;
 use sbor::rust::string::String;
