+++ radix-engine/src/system/client_api_impl.rs
@@ -9,7 +9,7 @@ use crate::kernel::actor::{Actor, ActorIdentifier};
 use crate::kernel::kernel::Kernel;
 use crate::kernel::kernel_api::KernelNodeApi;
 use crate::kernel::kernel_api::KernelSubstateApi;
-use crate::kernel::kernel_api::{Invokable, KernelInternalApi};
+use crate::kernel::kernel_api::{KernelInternalApi, KernelInvokeApi};
 use crate::kernel::module::KernelModule;
 use crate::kernel::module_mixer::KernelModuleMixer;
 use crate::system::kernel_modules::costing::FIXED_LOW_FEE;
